wip: review comments and cleanups
