+++ src/components/Tabs/types.ts
@@ -1,0 +1,16 @@
+import { ReactNode } from 'react';
+
+export interface TabItem {
+    id: string;
+    title: string;
+    content: ReactNode;
+}
+
+export interface TabsProps {
+    tabItems: TabItem[];
+}
+export interface TabNavProps {
+    items: TabItem[];
+    currentIndex: number;
+    onClick: (index: number) => void;
+}
