+++ src/components/transactions/history/ListItem.tsx
@@ -1,4 +1,11 @@
-import { TransactionInfo, TxType } from '@app/types/app-status.ts';
+import { useRef, useState } from 'react';
+import { AnimatePresence, useInView } from 'motion/react';
+import { useTranslation } from 'react-i18next';
+import { useAppConfigStore } from '@app/store/useAppConfigStore.ts';
+import { formatNumber, FormatPreset, truncateMiddle } from '@app/utils';
+import { BaseItemProps, HistoryListItemProps } from '../types.ts';
+import ItemExpand from './ExpandedItem';
+import ItemHover from './HoveredItem';
 import {
     ContentWrapper,
     ItemWrapper,
@@ -6,92 +13,13 @@ import {
     TitleWrapper,
     ValueChangeWrapper,
     ValueWrapper,
-    HoverWrapper,
-    ReplayButton,
-    ButtonWrapper,
-    FlexButton,
-    GemImage,
-    GemPill,
     CurrencyText,
-    InfoWrapper,
-    InfoItemWrapper,
 } from './ListItem.styles.ts';
-import { useRef, useState } from 'react';
-import { AnimatePresence, useInView } from 'motion/react';
-import { useTranslation } from 'react-i18next';
-
-import { useAppConfigStore } from '@app/store/useAppConfigStore.ts';
-import { formatNumber, FormatPreset, truncateMiddle } from '@app/utils';
-import { handleWinReplay } from '@app/store/useBlockchainVisualisationStore.ts';
-import { ReplaySVG } from '@app/assets/icons/replay.tsx';
-
-import gemImage from '@app/containers/main/Airdrop/AirdropGiftTracker/images/gem.png';
-import { GIFT_GEMS, useAirdropStore } from '@app/store/useAirdropStore.ts';
-import { useShareRewardStore } from '@app/store/useShareRewardStore.ts';
-
-interface HistoryListItemProps {
-    item: TransactionInfo;
-    showReplay?: boolean;
-    index: number;
-}
-
-interface BaseItemProps {
-    title: string;
-    type: TxType;
-    time: string;
-    value: string;
-    chip?: string;
-    onClick?: () => void;
-}
-
-function ItemExpand({ item }: { item: TransactionInfo }) {
-    const items = Object.keys(item).map((field) => (
-        <InfoItemWrapper key={field}>
-            <strong>{`${field}:`}</strong>
-            <span>{item[field]?.toString()}</span>
-        </InfoItemWrapper>
-    ));
-
-    return <InfoWrapper>{items}</InfoWrapper>;
-}
-function ItemHover({ item }: { item: TransactionInfo }) {
-    const { t } = useTranslation('sidebar', { useSuspense: false });
-    const sharingEnabled = useAppConfigStore((s) => s.sharing_enabled);
-    const referralQuestPoints = useAirdropStore((s) => s.referralQuestPoints);
-    const airdropTokens = useAirdropStore((s) => s.airdropTokens);
-    const { setShowModal, setItemData } = useShareRewardStore((s) => s);
-
-    const gemsValue = (referralQuestPoints?.pointsForClaimingReferral || GIFT_GEMS).toLocaleString();
-
-    const handleShareClick = () => {
-        setShowModal(true);
-        setItemData(item);
-    };
-    const isLoggedIn = !!airdropTokens;
-    const showShareButton = sharingEnabled && isLoggedIn;
-    return (
-        <HoverWrapper initial={{ opacity: 0 }} animate={{ opacity: 1 }} exit={{ opacity: 0 }}>
-            <ButtonWrapper initial={{ opacity: 0, y: 5 }} animate={{ opacity: 1, y: 0 }} exit={{ opacity: 0, y: 5 }}>
-                {showShareButton && (
-                    <FlexButton onClick={handleShareClick}>
-                        {t('share.history-item-button')}
-                        <GemPill>
-                            <span>{gemsValue}</span>
-                            <GemImage src={gemImage} alt="" />
-                        </GemPill>
-                    </FlexButton>
-                )}
-
-                <ReplayButton onClick={() => handleWinReplay(item)}>
-                    <ReplaySVG />
-                </ReplayButton>
-            </ButtonWrapper>
-        </HoverWrapper>
-    );
-}
 
 function BaseItem({ title, time, value, type, chip, onClick }: BaseItemProps) {
-    // TODO: check formatter - need to handle negative values
+    // note re. isPositiveValue:
+    // amounts in the tx response are always positive numbers but
+    // if the transaction type is 'sent' it must be displayed as a negative amount, with a leading `-`
     const isPositiveValue = type !== 'sent';
     const displayTitle = title.length > 30 ? truncateMiddle(title, 8) : title;
     return (
